# Fix string comparison with non-utf8 encodings