--- a/R/hash.R
+++ b/R/hash.R
@@ -3,8 +3,8 @@
 # 32 bit hashes. Thus, the size of the raw vector of hashes is 4 times
 # the size of the input.
 
-vec_hash <- function(x, rowwise = TRUE) {
-  .Call(vctrs_hash, x, rowwise)
+vec_hash <- function(x) {
+  .Call(vctrs_hash, x)
 }
 
 obj_hash <- function(x) {