--- a/src/equal.c
+++ b/src/equal.c
@@ -139,7 +139,11 @@ static int cpl_equal_scalar(const Rcomplex* x, const Rcomplex* y, bool na_equal)
 }
 
 static int chr_equal_scalar_impl(const SEXP x, const SEXP y) {
-  return !strcmp(Rf_translateCharUTF8(x), Rf_translateCharUTF8(y));
+  if (Rf_getCharCE(x) == Rf_getCharCE(y)) {
+    return x == y;
+  } else {
+    return !strcmp(Rf_translateCharUTF8(x), Rf_translateCharUTF8(y));
+  }
 }
 
 static int chr_equal_scalar(const SEXP* x, const SEXP* y, bool na_equal) {