--- a/src/hash.c
+++ b/src/hash.c
@@ -153,7 +153,7 @@ static uint32_t sexp_hash(SEXP x);
 uint32_t hash_object(SEXP x) {
   uint32_t hash = sexp_hash(x);
 
-  // Skip attribute check on CHARSXP, which might have attribute related
+  // Skip attribute check on CHARSXP, which might have an attribute related
   // to the global string pool (#555)
   if (TYPEOF(x) == CHARSXP) {
     return hash;