--- a/src/init.c
+++ b/src/init.c
@@ -14,7 +14,7 @@ extern SEXP vctrs_field_get(SEXP, SEXP);
 extern SEXP vctrs_field_set(SEXP, SEXP, SEXP);
 extern SEXP vctrs_fields(SEXP);
 extern SEXP vctrs_n_fields(SEXP);
-extern SEXP vctrs_hash(SEXP, SEXP );
+extern SEXP vctrs_hash(SEXP);
 extern SEXP vctrs_hash_object(SEXP);
 extern SEXP vctrs_equal_object(SEXP, SEXP, SEXP);
 extern SEXP vctrs_in(SEXP, SEXP);
@@ -83,7 +83,7 @@ static const R_CallMethodDef CallEntries[] = {
   {"vctrs_field_set",                  (DL_FUNC) &vctrs_field_set, 3},
   {"vctrs_fields",                     (DL_FUNC) &vctrs_fields, 1},
   {"vctrs_n_fields",                   (DL_FUNC) &vctrs_n_fields, 1},
-  {"vctrs_hash",                       (DL_FUNC) &vctrs_hash, 2},
+  {"vctrs_hash",                       (DL_FUNC) &vctrs_hash, 1},
   {"vctrs_hash_object",                (DL_FUNC) &vctrs_hash_object, 1},
   {"vctrs_equal_object",               (DL_FUNC) &vctrs_equal_object, 3},
   {"vctrs_in",                         (DL_FUNC) &vctrs_in, 2},