--- a/tests/testthat/test-dictionary.R
+++ b/tests/testthat/test-dictionary.R
@@ -108,32 +108,22 @@ test_that("unique functions take the equality proxy (#375)", {
   expect_identical(vec_match(tuple(2, 100), x), 2L)
 })
 
-test_that("vec_unique() can detect uniqueness with UTF-8 VS unknown encodings (#553)", {
-  utf8 <- "temp (\u00B0C)"
+test_that("vec_unique() can detect uniqueness with the same string in various encodings (#553)", {
+  utf8 <- "\u00B0C"
 
   unknown <- utf8
   Encoding(unknown) <- "unknown"
 
-  x <- c(unknown, utf8)
+  latin1 <- iconv(utf8, "UTF-8", "latin1")
+
+  x <- c(unknown, utf8, latin1)
 
   expect_equal(vec_unique(x), x[1])
   expect_equal(vec_unique(x), unique(x))
 })
 
-test_that("vec_unique() treats Latin1 vs unknown encoded strings as not equal", {
-  unknown <- "fa\xE7ile"
-
-  latin1 <- unknown
-  Encoding(latin1) <- "latin1"
-
-  x <- c(unknown, latin1)
-
-  expect_equal(vec_unique(x), x)
-  expect_equal(vec_unique(x), unique(x))
-})
-
 test_that("vec_unique() returns differently encoded strings in the order they appear", {
-  utf8 <- "temp (\u00B0C)"
+  utf8 <- "\u00B0C"
 
   unknown <- utf8
   Encoding(unknown) <- "unknown"
@@ -151,8 +141,7 @@ test_that("vec_unique() can determine uniqueness when the encoding is the same",
   latin1 <- unknown
   Encoding(latin1) <- "latin1"
 
-  utf8 <- unknown
-  Encoding(utf8) <- "UTF-8"
+  utf8 <- enc2utf8(latin1)
 
   x <- c(unknown, unknown)
   y <- c(latin1, latin1)
@@ -166,6 +155,25 @@ test_that("vec_unique() can determine uniqueness when the encoding is the same",
 
   expect_equal(vec_unique(z), z[1])
   expect_equal(vec_unique(z), unique(z))
+})
+
+test_that("vec_unique() works with bytes strings", {
+  utf8 <- "\u00B0C"
+
+  bytes <- utf8
+  Encoding(bytes) <- "bytes"
+
+  x <- c(bytes, bytes)
+
+  expect_equal(vec_unique(x), x[1])
+  expect_equal(vec_unique(x), unique(x))
+
+  y <- c(bytes, utf8)
+
+  # Error with base R
+  # unique(y)
+
+  expect_equal(vec_unique(y), y)
 })
 
 # matching ----------------------------------------------------------------