… tests/testthat/test-hash-hash.txt
… tests/testthat/test-hash.R