Related to #553 

CC @jennybc @paleolimbot

I did some investigation into this, and have a solution, but would really love another pair of eyes on this.

3 things:

- `chr_equal_scalar()` now supports comparing strings with different encodings. This is a part of the dictionary hashing, and also fixes these problems with `vec_equal()`.
- For the hashing, I added `hash_char()` to translate to UTF-8 before hashing a `CHARSXP`.
- Apparently, `CHARSXP` objects have attributes related to the global string pool. This caused me quite the headache, as I was unaware. I special cased them in `hash_object()`, so we don't check their attributes. I can't think of a better solution.

Base R skipping over `CHARSXP` attributes:
https://github.com/wch/r-source/blob/5a156a0865362bb8381dcd69ac335f5174a4f60c/src/main/identical.c#L99

Rationale for the `!strcmp(x,y)` approach used in `chr_equal_scalar_impl()`
https://stackoverflow.com/questions/37712672/and-in-differ-based-on-character-encoding

Feel free to modify/reject/redo this PR. Just wanted to get my thoughts down after investigating.

``` r
library(purrr)
library(vctrs)

bad_df_file <- tempfile(fileext = ".rds")
curl::curl_download(
  "https://gist.github.com/paleolimbot/ec9b62b758ae57a5b4669fa771fc40a0/raw/e96b55f54d68b1cb3877bb358b28b99dc8836ceb/bad_df.rds",
  bad_df_file
)
result_bad <- readRDS(bad_df_file)[["result"]]

names <- result_bad %>% 
  map(names) %>% 
  unlist()

# before
vec_unique(names)
#> [1] "Max Temp Flag"  "Min Temp (°C)"  "Min Temp Flag"  "Mean Temp (°C)"
#> [5] "Min Temp (°C)"  "Mean Temp (°C)"

# after
vec_unique(names)
#> [1] "Max Temp Flag"  "Min Temp (°C)"  "Min Temp Flag"  "Mean Temp (°C)"
```

```r
names[c(2, 6)]
#> [1] "Min Temp (°C)" "Min Temp (°C)"

# before
vec_equal(names[2], names[6])
#> [1] FALSE

# after
vec_equal(names[2], names[6])
#> [1] TRUE
```