@DavisVaughan 2019-08-30
@lionel- why do you use `intptr_t` over `int64_t` here?
https://github.com/r-lib/vctrs/blob/f5fe6e0913ec17a5079cc6ceaca98e0a37e90e5e/src/hash.c#L189

Should I be using `intptr_t` here?
https://github.com/r-lib/vctrs/pull/555/files#diff-292b11162bf53ed7cca27b794e4b092fR45