@jennybc 2019-08-30
Does vctrs already have a militant UTF-8 everywhere policy or anything like that?