@lionel- 2019-08-30
@DavisVaughan This is from https://github.com/r-lib/vctrs/commit/aadeac260d8f125aa0c3c61e6f38f665e442cfe2

I'm not sure what's best here but I guess coercing pointers to a 64 bit type should work almost everywhere. I don't think going through `intptr_t` matters though.