@hadley 2019-08-30
Would you mind also doing a quick benchmark in order to assess the performance cost?