@DavisVaughan 2019-08-30
Bad news is that the hashing got substantially slower, so `vec_unique()` is slow. Really it would be nice to know ahead of time whether a "translation" step is required or not. That would substantially speed things up, and is essentially what base R does. Here, if `useUTF8 = FALSE`, they don't convert.

https://github.com/wch/r-source/blob/2480f4d76dbdce30db615ee8c2cd43d655cb1241/src/main/unique.c#L178

Here is where they determine if that translation is required or not:
https://github.com/wch/r-source/blob/2480f4d76dbdce30db615ee8c2cd43d655cb1241/src/main/unique.c#L493

I'm not entirely sure how to refactor the code to allow for this, as right now its pretty hard to pass extra information on in a way that would feed it down to `hash_char()`.

``` r
library(vctrs)

# 10000 random words of size 10, append with utf-8 character
utf8 <- stringi::stri_rand_strings(10000, 10)
utf8 <- paste0(utf8, "\u00B0")

unknown <- utf8
Encoding(unknown) <- "unknown"
```

```r
# before
bench::mark(vec_unique(utf8), unique(utf8))
#> # A tibble: 2 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(utf8)    327µs    517µs     2014.     366KB     10.0
#> 2 unique(utf8)        263µs    354µs     2788.     245KB     14.6

# after
bench::mark(vec_unique(utf8), unique(utf8))
#> # A tibble: 2 x 6
#>   expression            min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>       <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(utf8)   12.8ms   13.7ms      72.4     366KB      0  
#> 2 unique(utf8)      272.9µs  345.9µs    2939.      245KB     12.2
```

```r
# before
bench::mark(vec_unique(unknown), unique(unknown))
#> # A tibble: 2 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(unknown)    410µs    504µs     2027.     348KB     12.4
#> 2 unique(unknown)        124µs    149µs     6529.     245KB     28.9

# after
bench::mark(vec_unique(unknown), unique(unknown))
#> # A tibble: 2 x 6
#>   expression               min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>          <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(unknown)   15.3ms   15.8ms      62.4     348KB     2.08
#> 2 unique(unknown)      136.1µs  183.4µs    5420.      245KB    24.7
```

```r
x <- c(utf8, unknown)

# before
# Error!

# after
bench::mark(vec_unique(x), unique(x))
#> # A tibble: 2 x 6
#>   expression         min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>    <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(x) 282.68ms  282.7ms      3.54     451KB     3.54
#> 2 unique(x)       5.56ms    6.2ms    158.       412KB     4.11
```

<sup>Created on 2019-08-30 by the [reprex package](https://reprex.tidyverse.org) (v0.2.1)</sup>

On the other hand `vec_equal()` is fairly close to `==` in speed, but they have access to a way to compare encodings that we don't have access to (`ENC_KNOWN()`), which does make it slower when the string pointers aren't equal (a common case). There are also some special cases with encoding that should be documented somewhere. This got a bit complicated, unfortunately, but essentially comparing a bytes encoded string + latin1/utf8 string is an error. We might be able to tweak the logic somehow to avoid this, I can't tell if it is a bug in base R or not.

<details>

``` r
library(vctrs)

# 10000 random words of size 10, append with utf-8 character
utf8 <- stringi::stri_rand_strings(10000, 10)
utf8 <- paste0(utf8, "\u00B0")

unknown <- utf8
Encoding(unknown) <- "unknown"

bytes <- unknown
Encoding(bytes) <- "bytes"

# Encodings are the same, and pointer comparison is always TRUE because they
# are the same
bench::mark(
  vec_equal(utf8, utf8), 
  utf8 == utf8
)
#> # A tibble: 2 x 6
#>   expression                 min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>            <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(utf8, utf8)   21.6µs   33.9µs    27780.    99.3KB     17.9
#> 2 utf8 == utf8            27.8µs   32.6µs    28920.    39.1KB     20.3

# Encodings are the same, and pointer comparison is always FALSE because they
# are all different, requiring an additional encoding check
no_match_utf8 <- rev(utf8)
bench::mark(
  vec_equal(utf8, no_match_utf8), 
  utf8 == no_match_utf8
)
#> # A tibble: 2 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                     <bch:> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(utf8, no_match_utf8) 51.1µs 61.4µs    16019.    39.1KB     9.94
#> 2 utf8 == no_match_utf8          35.7µs 38.8µs    24002.    39.1KB    16.4

# Encodings are different here
# - Pointer comparison will be false
# - Requires the encoding check, which returns FALSE
# - Then we do the expensive translate check
bench::mark(
  vec_equal(unknown, utf8), 
  unknown == utf8
)
#> # A tibble: 2 x 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(unknown, utf8)   2.52ms   2.68ms      366.    39.1KB     6.27
#> 2 unknown == utf8            2.44ms   2.57ms      388.    39.1KB     6.18

# bytes + unknown is allowed
bench::mark(
  vec_equal(bytes, unknown),
  bytes == unknown
)
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(bytes, unknown)   57.6µs   62.6µs    15213.    39.1KB     9.81
#> 2 bytes == unknown            38.6µs   44.1µs    21602.    39.1KB    16.8

# But base R throws an error for bytes+utf8/latin1
# - pointer comparison was FALSE
# - the encoding check returns FALSE
# - it tries to translate to utf8, but can't convert bytes
vec_equal(bytes, utf8) # or latin1
#> Error in vec_equal(bytes, utf8): translating strings with "bytes" encoding is not allowed
bytes == utf8
#> Error in eval(expr, envir, enclos): translating strings with "bytes" encoding is not allowed
```

</details>