@DavisVaughan 2019-08-31
I've come up with a clean new way to do this, speeds look good. I'll finalize it all Monday probably. Don't spend any time thinking about this