@DavisVaughan 2019-09-03
`vec_equal()` benchmarks

Setup:

```r
library(vctrs)
vec_equal_old <- vctrs2:::vec_equal

# 500000 random words of size 10, append with utf-8 character
utf8 <- stringi::stri_rand_strings(500000, 10)
utf8 <- paste0(utf8, "\u00B0")

rev_utf8 <- rev(utf8)

unknown <- utf8
Encoding(unknown) <- "unknown"

bytes <- utf8
Encoding(bytes) <- "bytes"

latin1 <- iconv(utf8, "UTF-8", "latin1")
```

Equal strings - very fast:

```r
bench::mark(
  old = vec_equal_old(utf8, utf8),
  new = vec_equal(utf8, utf8),
  base = utf8 == utf8,
  iterations = 500
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old        778.98µs 936.56µs      950.    1.96MB     56.4
#> 2 new          1.27ms   1.55ms      624.    1.96MB     24.6
#> 3 base         1.61ms   2.07ms      495.    1.91MB     18.5
```


All different - but same encoding. Requires an expensive-ish check to see if the encoding is the same. This is slightly slower than base R because we also check `na_equal` at each iteration, but without more refactoring I don't see any benefit to optimizing this.

```r
bench::mark(
  old = vec_equal_old(utf8, rev_utf8),
  new = vec_equal(utf8, rev_utf8),
  base = utf8 == rev_utf8,
  iterations = 500
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old        824.69µs 963.05µs     1021.    1.91MB    47.0 
#> 2 new          5.01ms   5.29ms      188.    1.91MB     8.22
#> 3 base         4.38ms   4.63ms      214.    1.91MB     8.91
```

All equal - but a translation is required to make it so.

```r
bench::mark(
  # wrong! old = vec_equal_old(utf8, unknown),
  new = vec_equal(utf8, unknown),
  base = utf8 == unknown,
  iterations = 50
)
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new           136ms    143ms      6.99    1.91MB     2.46
#> 2 base          132ms    137ms      7.26    1.91MB     3.11
```

Equivalent latin1/utf8

```r
vec_equal(latin1[1], utf8[1])
#> [1] TRUE

latin1[1] == utf8[1]
#> [1] TRUE
```

Base R error with bytes, but I just never attempt a translation if the strings are bytes.

```r
vec_equal(bytes[1], utf8[1])
#> [1] FALSE

bytes[1] == utf8[1]
#> Error in eval(expr, envir, enclos): translating strings with "bytes" encoding is not allowed
```