@DavisVaughan 2019-09-04
For completeness, `vec_unique()` benchmarks and special cases:

Setup:

```r
library(vctrs)
vec_unique_old <- vctrs2:::vec_unique

# 500000 random words of size 10, append with utf-8 character
utf8 <- stringi::stri_rand_strings(500000, 10)
utf8 <- paste0(utf8, "\u00B0")

unknown <- utf8
Encoding(unknown) <- "unknown"

bytes <- utf8
Encoding(bytes) <- "bytes"

latin1 <- iconv(utf8, "UTF-8", "latin1")
```

All values are unique, but same encoding

```r
bench::mark(
  old = vec_unique_old(utf8),
  new = vec_unique(utf8),
  base = unique(utf8),
  iterations = 20
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          31.1ms   37.9ms      27.2   15.64MB    22.2 
#> 2 new          37.1ms     46ms      21.6   15.65MB    11.6 
#> 3 base         44.5ms   62.2ms      16.9    9.72MB     7.23
```

All values are the same, with same encoding

```r
utf8_repeat <- rep(utf8[1], 500000)

bench::mark(
  old = vec_unique_old(utf8_repeat),
  new = vec_unique(utf8_repeat),
  base = unique(utf8_repeat),
  iterations = 100
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old           7.8ms   9.24ms     109.     5.91MB     19.2
#> 2 new          8.25ms   9.92ms     102.     5.91MB     13.9
#> 3 base         8.75ms  11.98ms      84.0    5.91MB     11.5
```

Notable case - All values are the same, all latin1 encoding. Big improvement over `unique()`, which tries to convert to UTF-8 here

```r
latin1_repeat <- iconv(utf8_repeat, "UTF-8", "latin1")

bench::mark(
  old = vec_unique_old(latin1_repeat),
  new = vec_unique(latin1_repeat),
  base = unique(latin1_repeat),
  iterations = 20
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 old          8.07ms   9.32ms    107.      5.91MB    18.9 
#> 2 new          9.31ms  10.32ms     96.8     5.91MB    24.2 
#> 3 base       141.92ms 149.58ms      6.67    5.91MB     3.59
```

Mixed case - UTF8 and unknown encodings together. First convert everything to UTF-8.

```r
utf8_unknown <- c(utf8, unknown)

length(utf8_unknown)
#> [1] 1000000
length(unique(utf8_unknown))
#> [1] 500000

# Mixed utf8 / unknown - translation required
bench::mark(
  # wrong! old = vec_unique_old(utf8_unknown),
  new = vec_unique(utf8_unknown),
  base = unique(utf8_unknown),
  iterations = 20
)
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new           447ms    450ms      2.22    21.5MB     9.44
#> 2 base          445ms    457ms      2.19    15.6MB     4.08
```

Can determine uniqueness among unknown / latin1 / utf8 encodings that all represent the same string.

```r
vec_unique(c(unknown[1], latin1[1], utf8[1]))
#> [1] "MkP6VHYFlx°"

unique(c(unknown[1], latin1[1], utf8[1]))
#> [1] "MkP6VHYFlx°"
```

_Special case_ - If any bytes string is detected, we never translate so the unknown/latin1/utf8 strings look different.

```r
vec_unique(c(bytes[1], unknown[1], latin1[1], utf8[1]))
#> [1] "MkP6VHYFlx\\xc2\\xb0" "MkP6VHYFlx°"        "MkP6VHYFlx°"       
#> [4] "MkP6VHYFlx°"

unique(c(bytes[1], unknown[1], latin1[1], utf8[1]))
#> [1] "MkP6VHYFlx\\xc2\\xb0" "MkP6VHYFlx°"        "MkP6VHYFlx°"       
#> [4] "MkP6VHYFlx°"
```